Fully remove ISequenceAdaptation
Fully removes the last traces of ISequenceAdaptation, along with a handful of related utility functions.
