+++ src/language-package/languages/seq-n/adaptation.ts
@@ -2,9 +2,9 @@ import { indentService } from "@codemirror/language";
 import { seqJsonToSeqn, seqnToSeqJson } from "@nasa-jpl/aerie-sequence-languages";
 import { EditorView } from "codemirror";
 import { debounce } from "lodash-es";
-import { outputLinter } from "../../interfaces/legacy";
 import type { LanguageAdaptation, NewAdaptationInterface, OutputLanguageAdaptation } from "../../interfaces/new-adaptation-interface";
 import { globals } from "./global-types";
+import { outputLinter } from "./output-linter";
 import { SeqLanguage, setupLanguageSupport } from "./seq-n";
 import { seqNBlockHighlighter, seqNHighlightBlock } from "./seq-n-highlighter";
 import { SeqNCommandInfoMapper } from "./seq-n-tree-utils";
